segregate ptz business rules
