+++ src/utils/ptz/ptz.types.ts
@@ -1,0 +1,9 @@
+export enum HousingQuotient {
+	'new' = 1,
+	'old',
+	'social',
+}
+
+export type Zone = 'A' | 'Abis' | 'B1' | 'B2' | 'C'
+
+export type NbrPeopleDigits = 1 | 2 | 3 | 4 | 5 | 6 | 7 | 8
